requirements.txt
customtkinter
groq
google-generativeai
python-dotenv
